requirements.txt
# Web Server & API
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1

# RAG & AI Framework
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.5

# Database & Vector Search
chromadb==0.5.0
pypdf==4.2.0

# Directory Loading & Processing
unstructured==0.14.4
extract-msg==0.48.7

# Utilities
pydantic==2.7.2
typing-extensions==4.12.1